Add functionality to draw 2-particle drip lines. Change all shorts to ints (shorts aren't used in the google coding convention as they vary in size between 32 and 64 bit machines).

+++ src/convertSymbolToZ.cpp
@@ -1,6 +1,6 @@
 #include "functions.h"
 
-unsigned short convertSymbolToZ(const std::string &el)
+int convertSymbolToZ(const std::string &el)
 {
        if (el == "n" ) return   0;
   else if (el == "H" ) return   1;
